+++ src/native/response/interface.ts
